django_filters, djcelery_email, celery usage with examples, flower monitoring tool in docker-compose, python3 downgrade because of missing celery support

+++ dockerX/todo/systemuser.py
@@ -5,6 +5,7 @@ import django; django.setup()
 from django.db import transaction
 models = importlib.import_module("%s.users.models" % os.environ['PROJECT'])
 
-with transaction.atomic():
-    models.User.objects.filter(email=os.environ['SYSTEM_EMAIL']).delete()
-    models.User.objects.create_superuser(os.environ['SYSTEM_EMAIL'], os.environ['SYSTEM_PASSWORD'])
+if not models.User.objects.filter(email=os.environ['SYSTEM_EMAIL']).exists():
+    with transaction.atomic():
+        models.User.objects.filter(email=os.environ['SYSTEM_EMAIL']).delete()
+        models.User.objects.create_superuser(os.environ['SYSTEM_EMAIL'], os.environ['SYSTEM_PASSWORD'])
